Initial support for value ranges
